delete WriterToClientThread.java refactored Constants
