+++ src/main/java/org/vshmaliukh/client_server/client/Client.java
@@ -1,6 +1,6 @@
-package org.vshmaliukh.client_server_model.client;
+package org.vshmaliukh.client_server.client;
 
-import org.vshmaliukh.client_server_model.ConstantsForClientServerModel;
+import org.vshmaliukh.constants.ConstantsForClientServerModel;
 
 import java.io.IOException;
 import java.io.PrintWriter;
